present calendar items
