+++ src/components/LifeEventsListView.vue
@@ -3,8 +3,8 @@
    <v-expansion-panels accordion multiple v-model="panel">
     
      <v-expansion-panel>
-       <v-expansion-panel-header>Next up</v-expansion-panel-header>
-       <v-expansion-panel-content>
+       <v-expansion-panel-title>Next up</v-expansion-panel-title>
+       <v-expansion-panel-text>
          <v-row dense>
            <v-col
                cols="12" xs="12" md="12"
@@ -13,11 +13,11 @@
              <LifeEvent :event="event" v-bind:key="event.googleId" v-on:reload="reload" />
            </v-col>
          </v-row>
-       </v-expansion-panel-content>
+       </v-expansion-panel-text>
      </v-expansion-panel>
      <v-expansion-panel>
-       <v-expansion-panel-header>The future</v-expansion-panel-header>
-       <v-expansion-panel-content>
+       <v-expansion-panel-title>The future</v-expansion-panel-title>
+       <v-expansion-panel-text>
          <v-row dense>
            <v-col
                cols="12" xs="12" md="12"
@@ -26,11 +26,11 @@
              <LifeEvent :event="event" v-bind:key="event.googleId" v-on:reload="reload" />
            </v-col>
          </v-row>
-       </v-expansion-panel-content>
+       </v-expansion-panel-text>
      </v-expansion-panel>
       <v-expansion-panel>
-       <v-expansion-panel-header>Past events</v-expansion-panel-header>
-       <v-expansion-panel-content>
+       <v-expansion-panel-title>Past events</v-expansion-panel-title>
+       <v-expansion-panel-text>
          <v-row dense>
            <v-col
                cols="12" xs="12" md="6"
@@ -39,60 +39,30 @@
              <LifeEvent :event="event" v-bind:key="event.googleId" v-on:reload="reload" />
            </v-col>
          </v-row>
-       </v-expansion-panel-content>
+       </v-expansion-panel-text>
      </v-expansion-panel>
    </v-expansion-panels>
  </v-container>
 </template>
 
 <script>
-import { compareAsc, addMonths } from 'date-fns'
 import LifeEvent from "@/components/LifeEvent.vue";
+import {mapState} from "pinia";
+import {useGoogleCalendarStore} from "@/stores/GoogleCalendarStore";
 
 export default {
   name: "LifeEventsListView",
   components: {LifeEvent},
-  props: ["events"],
   data: () => ({
     now: new Date(),
     panel: [1]
   }),
   computed: {
-    calendarId() {
-      return this.$store.state.calendarBackendId
-    },
-    oneMonthAhead() {
-      return addMonths(this.now, 1)
-    },
-    sortedEvents() {
-      return [...this.events]
-        ?.sort((a, b) => {
-          // first criteria: the end date
-          let dateSort = compareAsc(a.redZone.end, b.redZone.end)
-          if (dateSort !== 0) return dateSort
-          // second criteria: the title
-          if (a.title > b.title) return 1;
-          if (a.title < b.title) return -1;
-        })
-    },
-    pastEvents() {
-      return this.sortedEvents.filter((candidate) => candidate.redZone.end < this.now)
-    },
-    nextMonthEvents() {
-      return this.sortedEvents.filter((candidate) => candidate.redZone.end > this.now && candidate.redZone.end < this.oneMonthAhead)
-    },
-    futureEvents() {
-      return this.sortedEvents.filter((candidate) => candidate.redZone.end > this.oneMonthAhead)
-    }
+    ...mapState(useGoogleCalendarStore, ['sortedEvents', 'pastEvents', 'nextMonthEvents', 'futureEvents'])
   },
   methods: {
     reload() {
       this.$emit('reload')
-    }
-  },
-  watch: {
-    isLoading(newValue) {
-      this.$store.commit('setLoading', newValue)
     }
   }
 }
